Move BookList and BookForm to the containers folder
